decouple NftDetailAsset and NftDetailNote; move note into NftDetailPage
